Complete v1
This is a convenience program written to improve our
manual checking of PoS tagging done. This should serve
as a useful boilerplate to add in more functionality
as a user so wishes.

Next versions will cross check existing POS tagging libs
and ask the human only if there are discrepancies.

Written by: Alyssa Nah Xiao Ting and Ritesh Kumar

Co-Authored-By: alyssanxt <54824858+alyssanxt@users.noreply.github.com>