Reduce Human Prompts
Added a scoring system that calculates a weighted sum based on discrepancies between the original (stanford) tagging and other taggers.

Added a filtering to determine discrepencies wrt to a chosen threshold.

Added prompter that prompts human's checking of a SINGLE TERM.

Yet to check for bugs, yet to implement the weighted scoring, yet to introduce the pausing upon ctrl C feature.

Co-Authored-By: alyssanxt <54824858+alyssanxt@users.noreply.github.com>